refactor(frames): refactor airstack and implement get top 3 frames given fid
